+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 -  Add add/remove cli syntax for command `pcs stonith update-scsi-devices`
    ([rhbz#1992668])
 
+### Fixed
+- Do not unfence newly added devices on fenced cluster nodes ([rhbz#1991654])
+
+[rhbz#1991654]: https://bugzilla.redhat.com/show_bug.cgi?id=1991654
 [rhbz#1992668]: https://bugzilla.redhat.com/show_bug.cgi?id=1992668